Добавлено исключение для удаления категории с привязанными к ней событиями

+++ main/src/main/java/ru/practicum/service/AdminCategoryService.java
@@ -4,11 +4,16 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.stereotype.Service;
 import ru.practicum.dto.NewCategoryDto;
+import ru.practicum.exception.CategoryDeletionConflictException;
 import ru.practicum.exception.CategoryNotFoundException;
 import ru.practicum.exception.DuplicateCategoryException;
 import ru.practicum.mapper.CategoryMapper;
 import ru.practicum.model.Category;
+import ru.practicum.model.Event;
 import ru.practicum.repository.CategoryRepository;
+import ru.practicum.repository.EventRepository;
+
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -16,6 +21,8 @@ import ru.practicum.repository.CategoryRepository;
 public class AdminCategoryService {
 
     private final CategoryRepository categoryRepository;
+
+    private final EventRepository eventRepository;
 
     private final CategoryMapper categoryMapper;
 
@@ -32,6 +39,10 @@ public class AdminCategoryService {
     @Transactional
     public void deleteCategory(Long catId) {
         categoryRepository.findById(catId).orElseThrow(() -> new CategoryNotFoundException(catId.toString()));
+        List<Event> eventList = eventRepository.findAllByCategoryId(catId);
+        if (!eventList.isEmpty()) {
+            throw new CategoryDeletionConflictException("The category has linked events.");
+        }
         categoryRepository.deleteById(catId);
     }
